+++ src/main/kotlin/senior/joinu/candid/transpile/Utils.kt
@@ -254,13 +254,13 @@ fun transpileFunc(name: ClassName?, type: IDLType.Reference.Func, context: Trans
         getTypeSerForType(arg.type, requestTypeTable)
     }
     val typesSizeBytes =
-        Leb128.sizeUnsigned(argTypeSers.size.toLong()) + argTypeSers.map { it.calcTypeSizeBytes() }.sum()
+        Leb128.sizeUnsigned(argTypeSers.size) + argTypeSers.map { it.calcTypeSizeBytes() }.sum()
     val staticPayloadSize = MAGIC_PREFIX.size + requestTypeTable.sizeBytes() + typesSizeBytes
     val staticPayloadBuf = ByteBuffer.allocate(staticPayloadSize)
     staticPayloadBuf.put(MAGIC_PREFIX)
     requestTypeTable.serialize(staticPayloadBuf)
 
-    Leb128.writeUnsigned(staticPayloadBuf, argTypeSers.size.toLong())
+    Leb128.writeUnsigned(staticPayloadBuf, argTypeSers.size)
     argTypeSers.forEach { it.serType(staticPayloadBuf) }
 
     staticPayloadBuf.rewind()
